cleanup package ptree
