+++ pkg/gotkv/ptree/kv.go
@@ -3,7 +3,6 @@ package ptree
 import (
 	"bytes"
 	"context"
-	"fmt"
 
 	"github.com/brendoncarroll/go-state/cadata"
 	"github.com/gotvc/got/pkg/gdat"
@@ -16,6 +15,7 @@ type (
 	Entry = kvstreams.Entry
 )
 
+// MaxEntry returns the entry in span with the greatest (ordered last) key.
 func MaxEntry(ctx context.Context, s cadata.Store, x Root, span Span) (*Entry, error) {
 	op := gdat.NewOperator()
 	sr := NewStreamReader(s, &op, []Index{rootToIndex(x)})
@@ -96,48 +96,4 @@ func HasPrefix(ctx context.Context, s cadata.Store, x Root, prefix []byte) (bool
 		return false, nil
 	}
 	return true, nil
-}
-
-func DebugTree(s cadata.Store, x Root) {
-	max := x.Depth
-	op := gdat.NewOperator()
-	var debugTree func(Root)
-	debugTree = func(x Root) {
-		indent := ""
-		for i := 0; i < int(max-x.Depth); i++ {
-			indent += "  "
-		}
-		ctx := context.TODO()
-		sr := NewStreamReader(s, &op, []Index{{Ref: x.Ref, First: x.First}})
-		fmt.Printf("%sTREE NODE: %s %d\n", indent, x.Ref.CID.String(), x.Depth)
-		if x.Depth == 0 {
-			for {
-				var ent Entry
-				if err := sr.Next(ctx, &ent); err != nil {
-					if err == kvstreams.EOS {
-						break
-					}
-					panic(err)
-				}
-				fmt.Printf("%s ENTRY key=%q value=%q\n", indent, string(ent.Key), string(ent.Value))
-			}
-		} else {
-			for {
-				var ent Entry
-				if err := sr.Next(ctx, &ent); err != nil {
-					if err == kvstreams.EOS {
-						break
-					}
-					panic(err)
-				}
-				ref, err := gdat.ParseRef(ent.Value)
-				if err != nil {
-					panic(err)
-				}
-				fmt.Printf("%s INDEX first=%q -> ref=%s\n", indent, string(ent.Key), ref.CID.String())
-				debugTree(Root{Ref: *ref, First: ent.Key, Depth: x.Depth - 1})
-			}
-		}
-	}
-	debugTree(x)
 }
